improving tests + refactoring: test_embedding_retriever.py + test_keyword_retriever.py
