+++ integrations/supabase/src/haystack_integrations/components/retrievers/supabase/embedding_retriever.py
@@ -4,7 +4,7 @@
 
 from typing import Any, Literal
 
-from haystack import component, default_from_dict
+from haystack import component, default_from_dict, default_to_dict
 from haystack.document_stores.types import FilterPolicy
 
 from haystack_integrations.components.retrievers.pgvector import PgvectorEmbeddingRetriever
@@ -96,6 +96,22 @@ class SupabasePgvectorEmbeddingRetriever(PgvectorEmbeddingRetriever):
             top_k=top_k,
             vector_function=vector_function,
             filter_policy=filter_policy,
+        )
+
+    def to_dict(self) -> dict[str, Any]:
+        """
+        Serializes the component to a dictionary.
+
+        :returns:
+            Dictionary with serialized data.
+        """
+        return default_to_dict(
+            self,
+            filters=self.filters,
+            top_k=self.top_k,
+            vector_function=self.vector_function,
+            filter_policy=self.filter_policy.value,
+            document_store=self.document_store.to_dict(),
         )
 
     @classmethod
