+++ integrations/supabase/tests/test_keyword_retriever.py
@@ -9,171 +9,8 @@ from haystack.dataclasses import Document
 from haystack.document_stores.types import FilterPolicy
 from haystack.utils.auth import EnvVarSecret
 
-from haystack_integrations.components.retrievers.supabase import (
-    SupabasePgvectorEmbeddingRetriever,
-    SupabasePgvectorKeywordRetriever,
-)
+from haystack_integrations.components.retrievers.supabase import SupabasePgvectorKeywordRetriever
 from haystack_integrations.document_stores.supabase import SupabasePgvectorDocumentStore
-
-
-class TestEmbeddingRetriever:
-    def test_init_default(self, mock_store):
-        retriever = SupabasePgvectorEmbeddingRetriever(document_store=mock_store)
-        assert retriever.document_store == mock_store
-        assert retriever.filters == {}
-        assert retriever.top_k == 10
-        assert retriever.filter_policy == FilterPolicy.REPLACE
-        assert retriever.vector_function == mock_store.vector_function
-
-        retriever = SupabasePgvectorEmbeddingRetriever(document_store=mock_store, filter_policy="merge")
-        assert retriever.filter_policy == FilterPolicy.MERGE
-
-        with pytest.raises(ValueError):
-            SupabasePgvectorEmbeddingRetriever(document_store=mock_store, filter_policy="invalid")
-
-    def test_init(self, mock_store):
-        retriever = SupabasePgvectorEmbeddingRetriever(
-            document_store=mock_store, filters={"field": "value"}, top_k=5, vector_function="l2_distance"
-        )
-        assert retriever.document_store == mock_store
-        assert retriever.filters == {"field": "value"}
-        assert retriever.top_k == 5
-        assert retriever.filter_policy == FilterPolicy.REPLACE
-        assert retriever.vector_function == "l2_distance"
-
-    def test_init_invalid_document_store(self):
-        with pytest.raises(ValueError, match="must be an instance of SupabasePgvectorDocumentStore"):
-            SupabasePgvectorEmbeddingRetriever(document_store="not a store")
-
-    def test_to_dict(self, mock_store):
-        retriever = SupabasePgvectorEmbeddingRetriever(
-            document_store=mock_store, filters={"field": "value"}, top_k=5, vector_function="l2_distance"
-        )
-        res = retriever.to_dict()
-        t = (
-            "haystack_integrations.components.retrievers.supabase"
-            ".embedding_retriever.SupabasePgvectorEmbeddingRetriever"
-        )
-        assert res == {
-            "type": t,
-            "init_parameters": {
-                "document_store": {
-                    "type": (
-                        "haystack_integrations.document_stores.supabase.document_store.SupabasePgvectorDocumentStore"
-                    ),
-                    "init_parameters": {
-                        "connection_string": {"env_vars": ["SUPABASE_DB_URL"], "strict": True, "type": "env_var"},
-                        "create_extension": False,
-                        "schema_name": "public",
-                        "table_name": "haystack",
-                        "embedding_dimension": 768,
-                        "vector_type": "vector",
-                        "vector_function": "cosine_similarity",
-                        "recreate_table": True,
-                        "search_strategy": "exact_nearest_neighbor",
-                        "hnsw_recreate_index_if_exists": False,
-                        "language": "english",
-                        "hnsw_index_creation_kwargs": {},
-                        "hnsw_index_name": "haystack_hnsw_index",
-                        "hnsw_ef_search": None,
-                        "keyword_index_name": "haystack_keyword_index",
-                    },
-                },
-                "filters": {"field": "value"},
-                "top_k": 5,
-                "vector_function": "l2_distance",
-                "filter_policy": "replace",
-            },
-        }
-
-    @pytest.mark.usefixtures("patches_for_unit_tests")
-    def test_from_dict(self, monkeypatch):
-        monkeypatch.setenv("SUPABASE_DB_URL", "some-connection-string")
-        t = (
-            "haystack_integrations.components.retrievers.supabase"
-            ".embedding_retriever.SupabasePgvectorEmbeddingRetriever"
-        )
-        data = {
-            "type": t,
-            "init_parameters": {
-                "document_store": {
-                    "type": (
-                        "haystack_integrations.document_stores.supabase.document_store.SupabasePgvectorDocumentStore"
-                    ),
-                    "init_parameters": {
-                        "connection_string": {"env_vars": ["SUPABASE_DB_URL"], "strict": True, "type": "env_var"},
-                        "create_extension": False,
-                        "table_name": "haystack_test_to_dict",
-                        "embedding_dimension": 768,
-                        "vector_function": "cosine_similarity",
-                        "recreate_table": True,
-                        "search_strategy": "exact_nearest_neighbor",
-                        "hnsw_recreate_index_if_exists": False,
-                        "hnsw_index_creation_kwargs": {},
-                        "hnsw_index_name": "haystack_hnsw_index",
-                        "hnsw_ef_search": None,
-                        "keyword_index_name": "haystack_keyword_index",
-                    },
-                },
-                "filters": {"field": "value"},
-                "top_k": 5,
-                "vector_function": "l2_distance",
-                "filter_policy": "replace",
-            },
-        }
-
-        retriever = SupabasePgvectorEmbeddingRetriever.from_dict(data)
-        document_store = retriever.document_store
-
-        assert isinstance(document_store, SupabasePgvectorDocumentStore)
-        assert isinstance(document_store.connection_string, EnvVarSecret)
-        assert not document_store.create_extension
-        assert document_store.table_name == "haystack_test_to_dict"
-        assert document_store.embedding_dimension == 768
-        assert document_store.vector_function == "cosine_similarity"
-        assert document_store.recreate_table
-        assert document_store.search_strategy == "exact_nearest_neighbor"
-        assert not document_store.hnsw_recreate_index_if_exists
-        assert document_store.hnsw_index_creation_kwargs == {}
-        assert document_store.hnsw_index_name == "haystack_hnsw_index"
-        assert document_store.hnsw_ef_search is None
-        assert document_store.keyword_index_name == "haystack_keyword_index"
-
-        assert retriever.filters == {"field": "value"}
-        assert retriever.top_k == 5
-        assert retriever.filter_policy == FilterPolicy.REPLACE
-        assert retriever.vector_function == "l2_distance"
-
-    def test_run(self):
-        mock_store = Mock(spec=SupabasePgvectorDocumentStore)
-        mock_store.vector_function = "cosine_similarity"
-        doc = Document(content="Test doc", embedding=[0.1, 0.2])
-        mock_store._embedding_retrieval.return_value = [doc]
-
-        retriever = SupabasePgvectorEmbeddingRetriever(document_store=mock_store, vector_function="l2_distance")
-        res = retriever.run(query_embedding=[0.3, 0.5])
-
-        mock_store._embedding_retrieval.assert_called_once_with(
-            query_embedding=[0.3, 0.5], filters={}, top_k=10, vector_function="l2_distance"
-        )
-
-        assert res == {"documents": [doc]}
-
-    @pytest.mark.asyncio
-    async def test_run_async(self):
-        mock_store = Mock(spec=SupabasePgvectorDocumentStore)
-        mock_store.vector_function = "cosine_similarity"
-        doc = Document(content="Test doc", embedding=[0.1, 0.2])
-        mock_store._embedding_retrieval_async.return_value = [doc]
-
-        retriever = SupabasePgvectorEmbeddingRetriever(document_store=mock_store, vector_function="l2_distance")
-        res = await retriever.run_async(query_embedding=[0.3, 0.5])
-
-        mock_store._embedding_retrieval_async.assert_called_once_with(
-            query_embedding=[0.3, 0.5], filters={}, top_k=10, vector_function="l2_distance"
-        )
-
-        assert res == {"documents": [doc]}
 
 
 class TestKeywordRetriever:
@@ -377,13 +214,17 @@ class TestKeywordRetriever:
         doc = Document(content="Test doc", embedding=[0.1, 0.2])
         mock_store._keyword_retrieval.return_value = [doc]
 
-        retriever = SupabasePgvectorKeywordRetriever(
-            document_store=mock_store, filter_policy=FilterPolicy.MERGE, filters={"field": "value"}
-        )
-        res = retriever.run(query="test query", filters={"field2": "value2"})
+        init_filter = {"field": "meta.category", "operator": "==", "value": "news"}
+        runtime_filter = {"field": "meta.score", "operator": ">", "value": 0.5}
+        merged_filter = {"operator": "AND", "conditions": [init_filter, runtime_filter]}
 
-        mock_store.mock_calls[0].assert_called_once_with(
-            query="test query", filters={"field": "value", "field2": "value2"}, top_k=10
+        retriever = SupabasePgvectorKeywordRetriever(
+            document_store=mock_store, filter_policy=FilterPolicy.MERGE, filters=init_filter
+        )
+        res = retriever.run(query="test query", filters=runtime_filter)
+
+        mock_store._keyword_retrieval.assert_called_once_with(
+            query="test query", filters=merged_filter, top_k=10
         )
 
         assert res == {"documents": [doc]}
@@ -394,13 +235,64 @@ class TestKeywordRetriever:
         doc = Document(content="Test doc", embedding=[0.1, 0.2])
         mock_store._keyword_retrieval_async.return_value = [doc]
 
-        retriever = SupabasePgvectorKeywordRetriever(
-            document_store=mock_store, filter_policy=FilterPolicy.MERGE, filters={"field": "value"}
-        )
-        res = await retriever.run_async(query="test query", filters={"field2": "value2"})
+        init_filter = {"field": "meta.category", "operator": "==", "value": "news"}
+        runtime_filter = {"field": "meta.score", "operator": ">", "value": 0.5}
+        merged_filter = {"operator": "AND", "conditions": [init_filter, runtime_filter]}
 
-        mock_store.mock_calls[0].assert_called_once_with(
-            query="test query", filters={"field": "value", "field2": "value2"}, top_k=10
+        retriever = SupabasePgvectorKeywordRetriever(
+            document_store=mock_store, filter_policy=FilterPolicy.MERGE, filters=init_filter
+        )
+        res = await retriever.run_async(query="test query", filters=runtime_filter)
+
+        mock_store._keyword_retrieval_async.assert_called_once_with(
+            query="test query", filters=merged_filter, top_k=10
         )
 
         assert res == {"documents": [doc]}
+
+
+@pytest.mark.integration
+class TestKeywordRetrieval:
+    def test_keyword_retrieval(self, document_store: SupabasePgvectorDocumentStore):
+        docs = [
+            Document(content="The quick brown fox chased the dog", embedding=[0.1] * 768),
+            Document(content="The fox was brown", embedding=[0.1] * 768),
+            Document(content="The lazy dog", embedding=[0.1] * 768),
+            Document(content="fox fox fox", embedding=[0.1] * 768),
+        ]
+
+        document_store.write_documents(docs)
+
+        results = document_store._keyword_retrieval(query="fox", top_k=2)
+
+        assert len(results) == 2
+        for doc in results:
+            assert "fox" in doc.content
+        assert results[0].id == docs[-1].id
+        assert results[0].score > results[1].score
+
+    def test_keyword_retrieval_with_filters(self, document_store: SupabasePgvectorDocumentStore):
+        docs = [
+            Document(
+                content="The quick brown fox chased the dog",
+                embedding=[0.1] * 768,
+                meta={"meta_field": "right_value"},
+            ),
+            Document(content="The fox was brown", embedding=[0.1] * 768, meta={"meta_field": "right_value"}),
+            Document(content="The lazy dog", embedding=[0.1] * 768, meta={"meta_field": "right_value"}),
+            Document(content="fox fox fox", embedding=[0.1] * 768, meta={"meta_field": "wrong_value"}),
+        ]
+
+        document_store.write_documents(docs)
+
+        filters = {"field": "meta.meta_field", "operator": "==", "value": "right_value"}
+
+        results = document_store._keyword_retrieval(query="fox", top_k=3, filters=filters)
+        assert len(results) == 2
+        for doc in results:
+            assert "fox" in doc.content
+            assert doc.meta["meta_field"] == "right_value"
+
+    def test_empty_query(self, document_store: SupabasePgvectorDocumentStore):
+        with pytest.raises(ValueError):
+            document_store._keyword_retrieval(query="")
